refactor(rich-text-input): move components internal :)
