+++ src/components/internals/rich-text-body/dropdown.js
@@ -12,9 +12,9 @@ import {
   getButtonStyles,
   DropdownContainer,
   DropdownItem as StyledDropdownItem,
-} from './dropdown.styles';
-import messages from './messages';
-import { BLOCK_TAGS } from './utils/rules';
+} from '../../inputs/rich-text-input/dropdown.styles';
+import messages from '../../inputs/rich-text-input/messages';
+import { BLOCK_TAGS } from '../../inputs/rich-text-input/utils/rules';
 
 const propsToRemove = ['onClick'];
