+++ src/components/internals/rich-text-body/icons/index.js
