+++ src/components/internals/rich-text-body/index.js
@@ -1,0 +1,1 @@
+export { default } from './rich-text-body';
